massa_node

Module settings

Build here the default node settings from the configuration file toml

Structs

APISettings

API and server configuration, read from a file configuration.

BlockDumpSettings

BootstrapSettings

Bootstrap configuration.

CliSettings

ConsensusSettings

Consensus configuration Assumes thread_count >= 1, t0_millis >= 1, t0_millis % thread_count == 0

ExecutionSettings

FactorySettings

Factory settings

GrpcApiSettings

gRPC API settings.

GrpcSettings

gRPC settings the gRPC settings

LedgerSettings

LoggingSettings

MetricsSettings

NetworkSettings

PoolSettings

Pool configuration, read from a file configuration

ProtocolSettings

Protocol Configuration, read from toml user configuration file

SETTINGS

SelectionSettings

Settings

VersioningSettings